Add theme persistence and platform-specific favorite service
