+++ lib/main.dart
@@ -5,6 +5,9 @@ import 'core_scroll_behavior.dart';
 import 'pages/main_list_page.dart';
 import 'pages/detail_page.dart';
 import 'pages/settings_page.dart';
+import 'pages/list_page.dart';
+import 'pages/details_page.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 void main() {
   runApp(const AppRoot());
@@ -20,6 +23,26 @@ class AppRoot extends StatefulWidget {
 class _AppRootState extends State<AppRoot> {
   ThemeMode _mode = ThemeMode.light;
   int _tab = 0;
+
+  @override
+  void initState() {
+    super.initState();
+    _loadTheme();
+  }
+
+  Future<void> _loadTheme() async {
+    final prefs = await SharedPreferences.getInstance();
+    final saved = prefs.getInt('theme_mode');
+    if (saved != null && saved >= 0 && saved < ThemeMode.values.length) {
+      setState(() => _mode = ThemeMode.values[saved]);
+    }
+  }
+
+  void _setThemeMode(ThemeMode mode) async {
+    setState(() => _mode = mode);
+    final prefs = await SharedPreferences.getInstance();
+    await prefs.setInt('theme_mode', mode.index);
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -44,11 +67,13 @@ class _AppRootState extends State<AppRoot> {
                 const MainListPage(),
                 SettingsPage(
                   themeMode: _mode,
-                  onThemeModeChanged: (m) => setState(() => _mode = m),
+                  onThemeModeChanged: _setThemeMode,
                 ),
               ],
             ),
         '/detail': (_) => const DetailPage(),
+        '/list': (_) => const ListPage(),
+        '/details': (_) => const DetailsPage(),
       },
     );
   }
